+++ CHANGELOG.md
@@ -552,7 +552,7 @@
 * Added a new `display_name` field to `Pack` entity in graph.
 
 ## 1.21.0
-* Added the argument `--commited-only` to **pre-commit** to skip collecting on staged files.
+* Added the argument `--committed-only` to **pre-commit** to skip collecting on staged files.
 * Fixed an issue where the **pre-commit** command runs even in the case of deprecated or powershell integrations or scripts.
 * Fixed an issue where **validate** falsely failed with error `PB101` and `PB123` due to condition names discrepancy
 * Fixed an issue where the **modeling-rules test** command failed report and error when test data didn't exist.